// frontend/medicine-buddy-react/src/core/models/transaction/IGetDiseaseMedicineMappingTransactionModel.ts
export interface IGetDiseaseMedicineMappingTransactionModel {
  diseaseMedicineMappingId: number;
  diseaseId: number;
  diseaseTitle: string;
  diseaseDescription: string;
  diseaseIsActive: number;
  medicineId: number;
  medicineTitle: string;
  medicineDescription: string;
  medicineIsActive: boolean;
  transactionId: string;
}